+++ Sprites/Characters/Player.cs
@@ -10,6 +10,7 @@ namespace Project.Characters
     public class Player : Character, IInputReadable
     {
         public IInputReader InputReader { get; set; }
+        public Vector2 Velocity { get; set; }
         private Camera camera;
 
         public Player(Texture2D texture, IInputReader inputReader, Texture2D heartTexture, Camera camera) : base(texture, heartTexture)
@@ -22,10 +23,11 @@ namespace Project.Characters
             animationState.PlayAnimation(CharacterAnimation.IDLE);
 
             InputReader = inputReader;
-            Position = new Vector2(1, 1);
-            Speed = new Vector2(4, 3);
+            Position = Game1.mapManager.GetCurrentMap().MidOfMap;
+            Speed = new Vector2(3, 1.5f);
             Size = new Vector2(32, 32);
             Health = new Health(10, heartTexture);
+            Velocity = Vector2.Zero;
 
             this.camera = camera;
         }